requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# AI/ML - SAM3 and Vision
torch>=2.2.0
torchvision>=0.17.0
# Install transformers from main branch for SAM3 support
git+https://github.com/huggingface/transformers.git@main
accelerate>=1.2.0
safetensors>=0.4.3
huggingface-hub>=0.27.0

# Image/Video Processing
opencv-python-headless>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
av>=12.0.0

# Async utilities
aiohttp>=3.9.0
aiofiles>=23.2.0
httpx>=0.27.0

# Configuration
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Discord Integration
discord-webhook>=1.3.0

# Smart Home Integration
python-kasa>=0.6.2
pytapo>=3.3.54

# Audio (optional - may fail on some systems)
# simpleaudio>=1.0.4

# Utilities
python-dateutil>=2.8.0
structlog>=24.1.0
tenacity>=8.2.0

# For mask visualization
matplotlib>=3.8.0
